Glyphs for .loclBGR 3 Masters - Light, Regular, Black
Added additional .loclBGR glyph - Ka-cy.loclBGR, to match better the dynamics of the Latin, although it's not obligatory. The point is to reduce the number of horizontals in Bulgarian Cyrillic. That will further require the addition of Zhe-cy.loclBGR.

There are some inconsistent moments in Cyrillic- possibly issues in Russian/Squared/Standardized Glyphs:
- the "feet" of д-ц-щ share a metrics key of de-cy, although they have different width and that messes up the spacing. I decided to take the foot directly from de-cy and transfer it in loclBGR characters.

- there is a difference in ink traps in forms such as n-m-u, I transferred in loclBGR them exactly as they are (as of now).