Glyphs for .loclBGR 6 Masters
Two new glyphs added for loclBGR - iu-cy.loclBGR and iigrave-cy.loclBGR.
Minor changes and improvements to contours and metrics of the loclBGR glyphs.
Some further issues noted such as: difference in the positioning of the Cyrillic grave in Condensed masters, compared to the Latin grave. That doesn't make sense in Bulgarian context and leads to inconsistency.